+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 * Update :meth:`pyvcf.VcfFrame.plot_regplot` method to calculate and print summary statistics as well.
 * Add :meth:`pyvcf.VcfFrame.miss2ref` method.
 * Update :meth:`pymaf.MafFrame.plot_tmb` method to draw empty bars with warning when specified samples do not exist.
+* Update :meth:`pymaf.MafFrame.plot_waterfall` method to draw empty bars with warning when specified samples do not exist.
 
 0.17.0 (2021-07-08)
 -------------------